+++ apps/desktop/src/components/assistant-ui/thread-timestamp.ts
@@ -1,0 +1,39 @@
+const TIME_FMT = new Intl.DateTimeFormat(undefined, { hour: 'numeric', minute: '2-digit' })
+
+const SHORT_FMT = new Intl.DateTimeFormat(undefined, {
+  day: 'numeric',
+  hour: 'numeric',
+  minute: '2-digit',
+  month: 'short'
+})
+
+function startOfDay(d: Date): number {
+  return new Date(d.getFullYear(), d.getMonth(), d.getDate()).getTime()
+}
+
+export function formatMessageTimestamp(
+  value: Date | string | number | undefined,
+  labels: { today: (time: string) => string; yesterday: (time: string) => string }
+): string {
+  if (!value) {
+    return ''
+  }
+
+  const date = value instanceof Date ? value : new Date(value)
+
+  if (Number.isNaN(date.getTime())) {
+    return ''
+  }
+
+  const dayDelta = Math.round((startOfDay(new Date()) - startOfDay(date)) / 86_400_000)
+
+  if (dayDelta === 0) {
+    return labels.today(TIME_FMT.format(date))
+  }
+
+  if (dayDelta === 1) {
+    return labels.yesterday(TIME_FMT.format(date))
+  }
+
+  return SHORT_FMT.format(date)
+}
