+++ apps/desktop/src/components/assistant-ui/thread-types.ts
@@ -1,0 +1,4 @@
+export interface RestoreMessageTarget {
+  text: string
+  userOrdinal: number | null
+}
